nmocore/pyproject.toml
[project]
name = "nmocore"
version = "0.1.0"
description = "FastAPI REST API for NeuroMorpho.Org neuron morphology database"
requires-python = ">=3.10"
dependencies = [
    "psycopg2-binary>=2.9.3",
    "sqlalchemy>=1.4.37,<2",
    "fastapi>=0.78.0",
    "fastapi-pagination>=0.9.3",
    "uvicorn>=0.18.3",
    "paramiko>=2.11.0",
    "mysql-connector-python>=8.0.29",
    "SQLAlchemy-Utils>=0.38.2",
    "click>=8.1.3",
    "requests>=2.28.1",
    "redis>=4.3.4",
    "walrus>=0.9.2",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
]

[tool.setuptools]
packages = ["nmo"]
